fix get images

+++ src/pages/Login/Login.jsx
@@ -1,43 +1,54 @@
-import { useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
+import { useState } from 'react';
+import { useLocation } from 'react-router-dom';
 import Swal from 'sweetalert2';
 import withReactContent from 'sweetalert2-react-content';
 const mySwal = withReactContent(Swal);
-import { login } from './../../redux/slices/authSlice.js';
+import { useCombinedContexts } from './../../hooks/useCombineContexs.js';
 import { loginAPI, setAuthToken } from './../../services/authService.js';
 import './Login.css';
 
-export const Login = () => {
+export const Login = () => { 
+    const { setIsAuthenticated, login } = useCombinedContexts();
     const [credentials, setCredentials] = useState({ username: '', password: '' });
     const [error, setError] = useState('');
     const [loading, setLoading] = useState(false);
-    const navigate = useNavigate();
-    const dispatch = useDispatch();
-    const isAuthenticated = useSelector((state) => state.auth.isAuthenticated);
+    const location = useLocation();
+
+    const from = location.state?.from || '/dashboard';
 
     const handleLogin = async (e) => {
         e.preventDefault();
+        if (!credentials.username || !credentials.password) {
+            setError('Por favor, ingresa tu usuario y contraseña.');
+            mySwal.fire({
+                title: 'Error',
+                text: 'Por favor, ingresa tu usuario y contraseña.',
+                icon: 'warning',
+                confirmButtonText: 'Aceptar',
+            });
+            return;
+        }
         setLoading(true);
         setError('');
         try {
             const { access_token } = await loginAPI(credentials);
-            dispatch(login({ token: access_token }));
             setAuthToken(access_token);
             localStorage.setItem('access_token', access_token);
-            setLoading(false);
-            navigate('/dashboard');
+            setIsAuthenticated(true);
+            login(access_token, from)
+            
         } catch (err) {
-            console.error('Error en el inicio de sesión:', err);
-            setError(`Error en el inicio de sesión:', ${err}`)
-            alert(error);
+            setError(`Error en el inicio de sesión: ${err.message || 'Inténtalo nuevamente.'}`);
+            console.error(error);
+            setLoading(false);
+            mySwal.fire({
+                title: 'Error',
+                text: err.message || 'Ocurrió un error al iniciar sesión. Inténtalo nuevamente.',
+                icon: 'error',
+                confirmButtonText: 'Aceptar',
+            });
         }
     };
-    useEffect(()=>{
-        if (isAuthenticated) {
-            navigate('/dashboard')  
-        }        
-    })
 
     return (
         <main id="login" className="container">
@@ -45,37 +56,50 @@ export const Login = () => {
                 <h2 className="login__title">INGRESAR A MI CUENTA</h2>
                 <p className="login__subtitle">Para obtener novedades</p>
             </div>
-            <form className="login__form" action="" method="post" onSubmit={handleLogin}>
+            <form className="login__form" onSubmit={handleLogin}>
                 <div className="form__box--grid">
                     <label className="form__label" htmlFor="username">Usuario:</label>
-                    <input 
-                        className="form__input" 
-                        type="text" 
-                        name="username" 
-                        id="username" 
-                        placeholder="johndoe" 
-                        onChange={(e) => setCredentials({ ...credentials, username: e.target.value })}
+                    <input
+                        className="form__input"
+                        type="text"
+                        name="username"
+                        id="username"
+                        placeholder="johndoe"
+                        onChange={(e) =>
+                            setCredentials({ ...credentials, username: e.target.value }) &&
+                            setError('')
+                        }
                     />
                 </div>
                 <div className="form__box--grid">
                     <label className="form__label" htmlFor="password">Contraseña:</label>
-                    <input 
-                        className="form__input" 
-                        type="password" 
-                        name="password" 
-                        id="password" 
-                        placeholder="&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;" 
-                        onChange={(e) => setCredentials({ ...credentials, password: e.target.value })}
+                    <input
+                        className="form__input"
+                        type="password"
+                        name="password"
+                        id="password"
+                        placeholder="&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;&#9679;"
+                        onChange={(e) =>
+                            setCredentials({ ...credentials, password: e.target.value }) &&
+                            setError('')
+                        }
                     />
                 </div>
                 <div className="form__submission">
-                    <input className="form__btn btn btn--primary btn--large" type="submit" value={loading ? 'Ingresando...' : 'Ingresar'} onClick={() => { }}/>
+                    <input
+                        className="form__btn btn btn--primary btn--large"
+                        type="submit"
+                        value={loading ? 'Ingresando...' : 'Ingresar'}
+                        disabled={loading}
+                    />
                     <div className="form__remember">
                         <input type="checkbox" name="remember" id="" />
                         <label htmlFor="">Recordarme</label>
                     </div>
                 </div>
-                <a className="form__link" href="">Olvidé mi contraseña</a>
+                <a className="form__link" href="">
+                    Olvidé mi contraseña
+                </a>
             </form>
         </main>
     );
